Feat: folder, shared 페이지 반응형 처리 및 모바일 폴더 추가 버튼 추가
