+++ src/page/folder/FolderStyledComponents.js
@@ -17,7 +17,7 @@ export const Section = styled.div`
 `;
 
 export const ContentContainer = styled.div`
-  margin: 2rem auto 6rem;
+  margin: 6rem auto 9rem;
   display: flex;
   flex-direction: column;
   width: fit-content;
@@ -49,6 +49,12 @@ export const FolderContentContainer = styled(ContentContainer)`
   margin: 4rem auto 6rem;
   gap: 4rem;
   min-width: 106rem;
+  @media all and ${device.tablet} {
+    min-width: calc(100vw - 6.4rem);
+  }
+  @media all and ${device.mobile} {
+    min-width: calc(100vw - 6.4rem);
+  }
 
   .noLinkContainer {
     color: #000;
@@ -61,6 +67,66 @@ export const FolderContentContainer = styled(ContentContainer)`
   }
 `;
 
+export const AddLinkInputContainer = styled.div`
+  width: 100%;
+  min-width: 80rem;
+  display: flex;
+  padding: 1.6rem 2rem;
+  border-radius: 1.5rem;
+  background: #fff;
+  align-items: center;
+  box-sizing: border-box;
+  border: 1px solid var(--primary);
+  align-items: center;
+  @media all and ${device.tablet} {
+    min-width: calc(100vw - 6.4rem);
+  }
+  @media all and ${device.mobile} {
+    min-width: calc(100vw - 6.4rem);
+    padding: 0.8rem 1rem;
+  }
+
+  #linkImg {
+    margin-right: 1.2rem;
+    @media all and ${device.mobile} {
+      width: 1.6rem;
+    }
+  }
+
+  .inputContainer {
+    width: 100%;
+    height: 2.4rem;
+    padding-top: 0.5rem;
+  }
+
+  .inputContainer::placeholder {
+    color: var(--gray60);
+    font-size: 1.6rem;
+    font-weight: 400;
+    line-height: 2.4rem;
+    @media all and ${device.mobile} {
+      font-size: 1.4rem;
+    }
+  }
+
+  .linkAddBtn {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    padding: 1rem 1.6rem;
+    border-radius: 0.8rem;
+    background: linear-gradient(91deg, #6d6afe 0.12%, #6ae3fe 101.84%);
+    color: #f5f5f5;
+    font-size: 1.4rem;
+    font-weight: 600;
+    cursor: pointer;
+    box-sizing: content-box;
+    width: 8.1rem;
+    height: 1.7rem;
+    max-width: 8.1rem;
+  }
+`;
+
 export const SearchInputContainer = styled.div`
   width: 100%;
   display: flex;
@@ -70,17 +136,29 @@ export const SearchInputContainer = styled.div`
   background: #f5f5f5;
   align-items: center;
   box-sizing: border-box;
-
-  #searchContainer {
-    width: 100%;
-    height: 2.3rem;
+  align-items: center;
+  @media all and ${device.mobile} {
+    padding: 1.3rem 1.6rem;
   }
 
-  #searchContainer::placeholder {
+  .inputContainer {
+    width: 100%;
+    height: 2.3rem;
+    padding-top: 0.5rem;
+    @media all and ${device.mobile} {
+      height: 1.7rem;
+    }
+  }
+
+  .inputContainer::placeholder {
     color: #666;
     font-size: 1.6rem;
     font-weight: 400;
     line-height: 2.5rem;
+    @media all and ${device.mobile} {
+      font-size: 1.4rem;
+      line-height: 1.7rem;
+    }
   }
 
   #searchImg {
@@ -97,14 +175,21 @@ export const FolderContainer = styled.div`
   .folderBtnContainer {
     display: flex;
     gap: 0.8rem;
+    @media all and ${device.mobile} {
+      flex-wrap: wrap;
+      width: calc(100vw - 6.4rem);
+    }
   }
 
   .folderAddBtnContainer {
     display: flex;
     gap: 0.4rem;
     align-items: center;
-    jusitfy-content: center;
+    justify-content: center;
     cursor: pointer;
+    @media all and ${device.mobile} {
+      display: none;
+    }
 
     .folderAddTitle {
       color: #6d6afe;
@@ -133,6 +218,48 @@ export const FolderBtnItemContainer = styled.div`
   font-weight: 400;
 `;
 
+export const LinkHeaderContainer = styled.div`
+  display: flex;
+  width: 100%;
+  justify-content: space-between;
+  align-items: center;
+  @media all and ${device.mobile} {
+    flex-direction: column;
+    align-items: flex-start;
+    gap: 1.2rem;
+  }
+
+  .linkTitle {
+    color: #000;
+    font-size: 2.4rem;
+    font-weight: 600;
+    letter-spacing: -0.2px;
+  }
+`;
+
+export const LinkToolContainer = styled.div`
+  display: ${(props) => (props.display ? "none" : "flex")};
+  gap: 1.2rem;
+
+  .linkToolItemContainer {
+    display: flex;
+    gap: 1.2rem;
+    align-items: center;
+    cursor: pointer;
+
+    img {
+      width: 1.8rem;
+      height: 1.8rem;
+    }
+
+    .linkToolTitle {
+      color: var(--gray60);
+      font-size: 1.4rem;
+      font-weight: 600;
+    }
+  }
+`;
+
 export const CardContainer = styled.div`
   gap: 4rem;
   display: grid;
@@ -143,5 +270,36 @@ export const CardContainer = styled.div`
   }
   @media all and ${device.mobile} {
     grid-template-columns: repeat(1, 1fr);
+  }
+`;
+
+export const AddFloatingBtn = styled.div`
+  display: none;
+  @media all and ${device.mobile} {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    padding: 0.8rem 2.4rem;
+    border-radius: 2rem;
+    border: 1px solid #fff;
+    background: var(--primary);
+    width: fit-content;
+    height: 3.7rem;
+    box-sizing: border-box;
+    gap: 0.4rem;
+    position: fixed;
+    z-index: 5;
+    bottom: 10.1rem;
+    margin: 0 auto;
+    left: 50%;
+    transform: translate(-50%, 0);
+
+    .floatingBtnTitle {
+      color: var(--gray10);
+      text-align: center;
+      font-size: 1.6rem;
+      font-weight: 500;
+      letter-spacing: -0.3px;
+    }
   }
 `;
