+++ src/page/shared/Shared.js
@@ -2,83 +2,43 @@ import Card from "../../components/card/Card";
 import SearchImg from "../../assets/folder/img_search.png";
 import { useCallback, useEffect, useState } from "react";
 import useAsync from "../../hooks/useAsync";
-import { getFolder, getLinks } from "../../api/api";
+import { getSampleFolder } from "../../api/api";
 import {
-  AddLinkInputContainer,
   CardContainer,
   ContentContainer,
-  FolderBtnItemContainer,
-  FolderContainer,
   FolderContentContainer,
-  LinkHeaderContainer,
-  LinkToolContainer,
   SearchInputContainer,
   Section,
   Wrapper,
 } from "./SharedStyledComponents";
-import FolderAddIcon from "../../assets/folder/img_folderAdd.svg";
-import LinkAddIcon from "../../assets/shared/img_linkAdd.svg";
-import ShareIcon from "../../assets/shared/img_shareIcon.svg";
-import EditIcon from "../../assets/shared/img_editIcon.svg";
-import DeleteIcon from "../../assets/shared/img_deleteIcon.svg";
-
-const LinkToolArr = [
-  {
-    src: ShareIcon,
-    title: "공유",
-  },
-  {
-    src: EditIcon,
-    title: "이름 변경",
-  },
-  {
-    src: DeleteIcon,
-    title: "삭제",
-  },
-];
 
 const Folder = () => {
   const [cardData, setCardData] = useState([]);
-  const [folderData, setFolderData] = useState([]);
-  const [isSelectedFolder, setIsSelectedFolder] = useState({
-    id: 1,
-    title: "전체",
-  });
+  const [user, setUser] = useState();
+  const [folderName, setFolderName] = useState("");
 
-  const [isFolderLoading, isFolderError, getFolderAsync] = useAsync(getFolder);
-  const [isLinkLoading, isLinkError, getLinksAsync] = useAsync(getLinks);
+  const [isLoading, error, getSampleFolderAsync] = useAsync(getSampleFolder);
 
   const handleFolder = useCallback(async () => {
-    const result = await getFolderAsync();
+    const result = await getSampleFolderAsync();
     if (!result) return;
 
-    const { data } = result;
+    const { folder } = result;
 
-    setFolderData(data);
-  }, [getFolderAsync]);
+    setCardData(folder.links);
+    setUser(folder.owner);
+    setFolderName(folder.name);
+  }, [getSampleFolderAsync]);
 
   useEffect(() => {
     handleFolder();
   }, [handleFolder]);
 
-  const handleLinks = useCallback(async () => {
-    const result = await getLinksAsync(isSelectedFolder.id);
-    if (!result) return;
-
-    const { data } = result;
-
-    setCardData(data);
-  }, [getLinksAsync, isSelectedFolder]);
-
-  useEffect(() => {
-    handleLinks();
-  }, [handleLinks]);
-
-  if (isFolderLoading && isLinkLoading) {
+  if (isLoading) {
     return <div>화면을 불러오는 중입니다.</div>;
   }
 
-  if (isFolderError && isLinkError) {
+  if (error) {
     return <div>문제가 발생했습니다.</div>;
   }
 
@@ -86,14 +46,13 @@ const Folder = () => {
     <Wrapper>
       <Section>
         <ContentContainer>
-          <AddLinkInputContainer>
-            <img src={LinkAddIcon} alt="linkImg" id="linkImg" />
-            <input
-              className="inputContainer"
-              placeholder="링크를 추가해 보세요"
-            ></input>
-            <button className="linkAddBtn">추가하기</button>
-          </AddLinkInputContainer>
+          <img
+            id="userProfile"
+            src={user?.profileImageSource}
+            alt="profileImg"
+          />
+          <div id="userName">{user?.name}</div>
+          <div id="folderName">{folderName}</div>
         </ContentContainer>
       </Section>
       <Section bg="#fff">
@@ -101,78 +60,15 @@ const Folder = () => {
           <SearchInputContainer>
             <img src={SearchImg} alt="searchImg" id="searchImg" />
             <input
-              className="inputContainer"
+              id="searchContainer"
               placeholder="링크를 검색해보세요."
             ></input>
           </SearchInputContainer>
-
-          {cardData.length > 0 ? (
-            <>
-              <FolderContainer>
-                <div className="folderBtnContainer">
-                  <FolderBtnItemContainer
-                    isSelected={isSelectedFolder.id === 1}
-                    onClick={() =>
-                      setIsSelectedFolder({
-                        id: 1,
-                        title: "전체",
-                      })
-                    }
-                  >
-                    전체
-                  </FolderBtnItemContainer>
-
-                  {folderData?.map((e) => {
-                    return (
-                      <FolderBtnItemContainer
-                        key={e.id}
-                        isSelected={e.id === isSelectedFolder}
-                        onClick={() =>
-                          setIsSelectedFolder({
-                            id: e.id,
-                            title: e.name,
-                          })
-                        }
-                      >
-                        {e.name}
-                      </FolderBtnItemContainer>
-                    );
-                  })}
-                </div>
-
-                <div className="folderAddBtnContainer">
-                  <div className="folderAddTitle">폴더 추가</div>
-                  <img
-                    src={FolderAddIcon}
-                    className="folderAddIcon"
-                    alt="folderAddIcon"
-                  />
-                </div>
-              </FolderContainer>
-              <LinkHeaderContainer>
-                <div className="linkTitle">{isSelectedFolder.title}</div>
-
-                <LinkToolContainer>
-                  {LinkToolArr.map((e, index) => {
-                    return (
-                      <div className="linkToolItemContainer" key={index}>
-                        <img src={e.src} alt={e.title} />
-                        <div className="linkToolTitle">{e.title}</div>
-                      </div>
-                    );
-                  })}
-                </LinkToolContainer>
-              </LinkHeaderContainer>
-
-              <CardContainer>
-                {cardData?.map((e) => {
-                  return <Card key={e.id} cardData={e} />;
-                })}
-              </CardContainer>
-            </>
-          ) : (
-            <div className="noLinkContainer">저장된 링크가 없습니다</div>
-          )}
+          <CardContainer>
+            {cardData?.map((e) => {
+              return <Card key={e.id} cardData={e} />;
+            })}
+          </CardContainer>
         </FolderContentContainer>
       </Section>
     </Wrapper>
